+++ src/knowledge/knowledge-self.cpp
@@ -55,9 +55,9 @@ static void dump_yourself(player_type *creature_ptr, FILE *fff)
         return;
 
     char temp[80 * 10];
-    shape_buffer(race_explanations[creature_ptr->prace], 78, temp, sizeof(temp));
+    shape_buffer(race_explanations[static_cast<int>(creature_ptr->prace)], 78, temp, sizeof(temp));
     fprintf(fff, "\n\n");
-    fprintf(fff, _("種族: %s\n", "Race: %s\n"), race_info[creature_ptr->prace].title);
+    fprintf(fff, _("種族: %s\n", "Race: %s\n"), race_info[static_cast<int>(creature_ptr->prace)].title);
     concptr t = temp;
 
     for (int i = 0; i < 10; i++) {
