+++ src/player-status/player-basic-statistics.cpp
@@ -49,7 +49,7 @@ s16b PlayerBasicStatistics::race_value()
     if (this->owner_ptr->mimic_form)
         tmp_rp_ptr = &mimic_info[this->owner_ptr->mimic_form];
     else
-        tmp_rp_ptr = &race_info[this->owner_ptr->prace];
+        tmp_rp_ptr = &race_info[static_cast<int>(this->owner_ptr->prace)];
 
     return tmp_rp_ptr->r_adj[this->ability_type];
 }
