+++ src/view/display-store.cpp
@@ -172,7 +172,7 @@ void display_store(player_type *player_ptr)
 
     concptr store_name = f_info[cur_store_feat].name.c_str();
     concptr owner_name = (ot_ptr->owner_name);
-    concptr race_name = race_info[ot_ptr->owner_race].title;
+    concptr race_name = race_info[static_cast<int>(ot_ptr->owner_race)].title;
     char buf[80];
     sprintf(buf, "%s (%s)", owner_name, race_name);
     put_str(buf, 3, 10);
